+++ jax/random.py
@@ -144,7 +144,6 @@ from jax._src.random import (
   cauchy as cauchy,
   chisquare as chisquare,
   choice as choice,
-  default_prng_impl as _deprecated_default_prng_impl,
   dirichlet as dirichlet,
   double_sided_maxwell as double_sided_maxwell,
   exponential as exponential,
@@ -183,27 +182,7 @@ from jax._src.random import (
   wrap_key_data as wrap_key_data,
 )
 
-from jax._src.prng import (
-  threefry_2x32 as _deprecated_threefry_2x32,
-  threefry2x32_p as _deprecated_threefry2x32_p,
-)
-
 _deprecations = {
-    # Added October 18, 2023
-    "threefry_2x32": (  # Note: this has been raising a FutureWarning since 2021
-        "jax.random.threefry_2x32 is deprecated. Use jax.extend.random.threefry_2x32.",
-        _deprecated_threefry_2x32,
-    ),
-    "threefry2x32_p": (
-        "jax.random.threefry2x32_p is deprecated. Use jax.extend.random.threefry2x32_p.",
-        _deprecated_threefry2x32_p,
-    ),
-    # Added October 19. 2023
-    "default_prng_impl": (
-        "jax.random.default_prng_impl is deprecated. Typical uses can be replaced by "
-        "jax.random.key_impl(key), jax.eval_shape(jax.random.key, 0).dtype, or similar.",
-        _deprecated_default_prng_impl,
-    ),
     # Added November 6, 2023; but has been raising a FutureWarning since JAX 0.1.66
     "shuffle": (
         "jax.random.shuffle is deprecated. Use jax.random.permutation with independent=True.",
@@ -213,10 +192,7 @@ _deprecations = {
 
 import typing
 if typing.TYPE_CHECKING:
-  default_prng_impl = _deprecated_default_prng_impl
   shuffle = _deprecated_shuffle
-  threefry_2x32 = _deprecated_threefry_2x32
-  threefry2x32_p = _deprecated_threefry2x32_p
 else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr
   __getattr__ = _deprecation_getattr(__name__, _deprecations)